refactor: Rename api folder.
