+++ app/utils/db/sqlite/command-listener.ts
@@ -1,8 +1,8 @@
-import { apiCommandBus } from '#utils/api/command-bus.server.js';
+import { commandBus } from '#utils/cqrs/command-bus.server';
 import { createChampionshipCommand } from './commands/create-championship';
 
 export function registerSqlite() {
-  apiCommandBus.on(({ name, payload }) => {
+  commandBus.on(({ name, payload }) => {
     switch (name) {
       case 'CREATE_CHAMPIONSHIP':
         createChampionshipCommand(payload);
