+++ app/utils/db/sqlite/commands/create-championship.ts
@@ -1,4 +1,4 @@
-import type { CreateChampionshipCommand } from '#utils/api/commands';
+import type { CreateChampionshipCommand } from '#utils/cqrs/commands';
 import { db } from '#utils/db.server';
 
 export async function createChampionshipCommand(
